#13 업적 목록 엔티티 추가
업적 목록 엔티티 정의, crud 서비스 코드 추가

+++ src/main/java/com/polychat/polychatbe/achievement/query/repository/AchievementSearchRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface AchievementSearchRepository {
 
     long countAchievement();
-    AchievementResponseDTO findAchievementById(Integer achievementId);
+    AchievementResponseDTO findAchievementById(Long achievementId);
     List<AchievementResponseDTO> findAllAchievement();
 
 }
